Core: move I/O-related functions from lib/common/utils.c to new io.c
Specifically:
 crm_build_path()
 generate_series_filename()
 get_last_sequence()
 write_last_sequence()
 crm_is_writable()
