+++ include/crm/common/Makefile.am
@@ -21,7 +21,7 @@ MAINTAINERCLEANFILES = Makefile.in
 headerdir=$(pkgincludedir)/crm/common
 
 header_HEADERS = xml.h ipc.h util.h iso8601.h mainloop.h logging.h
-noinst_HEADERS = ipcs.h
+noinst_HEADERS = io.h ipcs.h
 if BUILD_CIBSECRETS
 noinst_HEADERS += cib_secrets.h
 endif
